extracted implementation part of `expr.h` to `expr.c` and others.
extracted to `expr.c`, `decl.c`, `stmts.c`, and `identifier.c`.
